nail-salon/backend/requirements.txt
django==4.2
djangorestframework==3.14
django-cors-headers==4.0